refactor: move findBooksBySearchString
- move findBooksBySearchString to the search lib to lighten to load in the actions lib, and as it's a better fit in search
