+++ src/lib/actions/book.test.ts
@@ -1,7 +1,6 @@
 import {
   buildAuthorsInput,
   buildPublisherInput,
-  findBooksBySearchString,
   getBook,
   getBooks,
   upsertBook,
@@ -204,71 +203,6 @@ describe('book actions', () => {
           startCursor: book2.id.toString(),
         },
       });
-    });
-  });
-
-  describe('findBooksBySearchString', () => {
-    it('should find books that contain "Book"', async () => {
-      prismaMock.book.findMany.mockResolvedValue([book1, book2, book3]);
-
-      const result = await findBooksBySearchString('Book');
-
-      expect(prismaMock.book.findMany).toHaveBeenCalledWith({
-        include: {
-          authors: true,
-          format: true,
-          genre: true,
-          publisher: true,
-        },
-        where: {
-          OR: [
-            { authors: { some: { name: { search: 'Book' } } } },
-            { title: { search: 'Book' } },
-          ],
-        },
-      });
-      expect(result).toEqual([book1, book2, book3]);
-    });
-
-    it('should process multiple words', async () => {
-      prismaMock.book.findMany.mockResolvedValue([book2]);
-
-      const result = await findBooksBySearchString('the other book');
-
-      expect(prismaMock.book.findMany).toHaveBeenCalledWith({
-        include: {
-          authors: true,
-          format: true,
-          genre: true,
-          publisher: true,
-        },
-        where: {
-          OR: [
-            { authors: { some: { name: { search: 'the & other & book' } } } },
-            { title: { search: 'the & other & book' } },
-          ],
-        },
-      });
-      expect(result).toEqual([book2]);
-    });
-
-    it('should process ISBN13', async () => {
-      prismaMock.book.findMany.mockResolvedValue([book2]);
-
-      const result = await findBooksBySearchString('9780765376671');
-
-      expect(prismaMock.book.findMany).toHaveBeenCalledWith({
-        include: {
-          authors: true,
-          format: true,
-          genre: true,
-          publisher: true,
-        },
-        where: {
-          OR: [{ isbn13: { equals: 9780765376671 } }],
-        },
-      });
-      expect(result).toEqual([book2]);
     });
   });
