+++ src/lib/actions/book.ts
@@ -13,7 +13,6 @@ import BookHydrated from '@/types/BookHydrated';
 import { Book, Prisma, ProductType } from '@prisma/client';
 import { serializeBookSource } from '@/lib/serializers/book-source';
 import NegativeBookQuantityError from '@/lib/errors/NegativeBookQuantityError';
-import _ from 'lodash';
 
 export async function buildAuthorsInput(
   tx: Prisma.TransactionClient,
@@ -250,43 +249,6 @@ export async function getBooks({
     books,
     pageInfo,
   };
-}
-
-export async function findBooksBySearchString(
-  searchString: string,
-): Promise<Array<BookHydrated>> {
-  let whereClauses: Prisma.BookWhereInput[];
-  const searchStringNumber = _.toNumber(searchString);
-  if (_.isFinite(searchStringNumber)) {
-    whereClauses = [{ isbn13: { equals: searchStringNumber } }];
-  } else {
-    const search = searchString.split(' ').join(' & ');
-    whereClauses = [
-      { authors: { some: { name: { search } } } },
-      { title: { search } },
-    ];
-  }
-
-  const rawBooks = await prisma.book.findMany({
-    include: {
-      authors: true,
-      format: true,
-      genre: true,
-      publisher: true,
-    },
-    where: {
-      OR: whereClauses,
-    },
-  });
-
-  const books = rawBooks.map((book) => ({
-    ...book,
-    publisher: serializeBookSource(book.publisher),
-  }));
-
-  logger.trace('books found: %j', books);
-
-  return books;
 }
 
 export async function getBook(
